dockers/fed_reserve_features/requirements.txt
pandas==1.3.4
full-fred==0.0.9a3
pytrends==4.9.0
kafka-python==2.0.2
requests-html==0.10.0
requests-oauthlib==1.3.1
psycopg2-binary==2.9.6
sqlalchemy==1.4.46
google-cloud-storage==2.5.0